backend/requirements.txt
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==2.1.5
mistune==3.0.2
mongoengine==0.29.1
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
ruff==0.6.7
six==1.16.0
Werkzeug==3.0.4